services/knowledgebase/requirements.txt
fastapi[standard]

sqlalchemy
asyncpg
python-jose
bcrypt
pydantic
redis
aioboto3
aiokafka
neo4j
motor
beanie
openai

PyPDF2
python-docx
python-pptx
pdf2image
pytesseract